MSME_bot2/requirements.txt
fastapi
uvicorn
openai
python-dotenv
twilio
